+++ src/modules/core/Color.h
@@ -20,7 +20,7 @@ union RGBA {
 
 class Color {
 public:
-	static const unsigned int magnitude = 255;
+	static const uint32_t magnitude = 255;
 	static constexpr float magnitudef = 255.0f;
 	static constexpr float scaleFactor = 0.7f;
 	static const glm::vec4
@@ -52,7 +52,9 @@ public:
 		LightBrown,
 		DarkBrown;
 
+	static float getDistance(uint32_t rgba, uint32_t rgba2);
 	static float getDistance(const glm::vec4& color, float hue, float saturation, float brightness);
+	static float getDistance(uint32_t color, float hue, float saturation, float brightness);
 
 	/**
 	 * @brief Get the nearest matching color index from the list
@@ -60,7 +62,7 @@ public:
 	 * @return index in the colors vector or the first entry if non was found, or @c -1 on error
 	 */
 	template<class T>
-	static int getClosestMatch(const glm::vec4& color, const T& colors) {
+	static int getClosestMatch(const glm::vec4& color, const T& colors, float *distance = nullptr) {
 		if (colors.empty()) {
 			return -1;
 		}
@@ -80,28 +82,32 @@ public:
 				minIndex = (int)i;
 			}
 		}
+		if (distance) {
+			*distance = minDistance;
+		}
 		return minIndex;
 	}
 
-	static glm::vec4 fromRGB(const unsigned int rgbInt, const float a = 1.0f);
-	static glm::vec4 fromRGBA(const unsigned int rgbaInt);
-	static glm::u8vec4 toRGBA(const unsigned int rgbaInt);
-	static glm::vec4 fromARGB(const unsigned int argbInt);
-	static glm::vec4 fromRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
+	static glm::vec4 fromRGB(const uint32_t rgbInt, const float a = 1.0f);
+	static glm::vec4 fromRGBA(const uint32_t rgbaInt);
+	static glm::u8vec4 toRGBA(const uint32_t rgbaInt);
+	static glm::vec4 fromARGB(const uint32_t argbInt);
+	static glm::vec4 fromRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
 	static glm::vec4 fromHSB(const float hue, const float saturation, const float brightness, const float alpha = 1.0f);
 	static glm::vec4 fromHex(const char* hex);
-	static core::String toHex(const unsigned int rgba, bool hashPrefix = true);
+	static core::String toHex(const uint32_t rgba, bool hashPrefix = true);
 
-	static unsigned int getRGB(const glm::vec4&);
-	static unsigned int getRGBA(const glm::vec4&);
-	static unsigned int getRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a);
+	static uint32_t getRGB(const glm::vec4&);
+	static uint32_t getRGBA(const glm::vec4&);
+	static uint32_t getRGBA(uint8_t r, uint8_t g, uint8_t b, uint8_t a = 255);
 	static glm::u8vec4 getRGBAVec(const glm::vec4&);
-	static unsigned int getBGRA(const glm::vec4& color);
+	static uint32_t getBGRA(const glm::vec4& color);
 	static void getHSB(const glm::vec4&, float& hue, float& saturation, float& brightness);
 
 	static glm::vec3 gray(const glm::vec3&);
 	static glm::vec4 gray(const glm::vec4&);
 	static glm::vec4 alpha(const glm::vec4&, float alpha);
+	static uint32_t alpha(const uint32_t rgba, uint8_t alpha);
 	static float brightness(const glm::vec4&);
 	static float intensity(const glm::vec4&);